requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==7.1.2
cryptography==44.0.3
dnspython==2.6.1
email-validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.61.2
frozenlist==1.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jose==1.0.0
multidict==6.1.0
passlib==1.7.4
propcache==0.2.0
pycparser==2.22
pydantic==1.7.4
pydantic-core==2.27.2
PyJWT==2.9.0
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.13.6
typing-extensions==4.13.2
urllib3==2.2.3
uvicorn==0.13.4
yarl==1.15.2
